refactor: organize all hooks into consistent directory structure - Move all single-file hooks into directories with index.ts files - Update README to reflect new directory structure - Add documentation for newly organized hooks - Maintain consistent directory-per-hook pattern - Include all student-specific and utility hooks in documentation
